+++ tv/pages/room.tsx
@@ -1,9 +1,9 @@
 import Head from 'next/head';
-import Image from 'next/image';
 import { getBlogs } from '../api/queries/getBlogs';
 import { GetStaticProps } from 'next';
 import { Blog } from '../interfaces';
 import RoomDisplay from '../components/RoomDisplay';
+import Footer from '../components/Footer';
 
 type RoomProps = {
   blogs: Blog[];
@@ -29,17 +29,8 @@ export default function Room(props: RoomProps) {
         </div>
       </main>
 
-      <footer className="footer">
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{' '}
-          <span className="logo">
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
+      <footer>
+        <Footer />
       </footer>
     </div>
   );
