+++ tv/pages/speaker.tsx
@@ -1,6 +1,7 @@
 import Head from 'next/head';
 import Image from 'next/image';
 import banner1 from '../public/banner1.png';
+import Footer from '../components/Footer';
 
 export default function Speaker() {
   return (
@@ -22,17 +23,8 @@ export default function Speaker() {
         </div>
       </main>
 
-      <footer className="footer">
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{' '}
-          <span className="logo">
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
+      <footer>
+        <Footer />
       </footer>
     </div>
   );
