+++ tv/styles/layout/footer.css
@@ -54,6 +54,7 @@
         display: block;
         justify-content: space-evenly;
         align-items: center;
+        text-align: center;
         flex-grow: 1;
         padding: 15px;
       }
